combine provider.go and pool.go into driver.go
google move defaults to options
