+++ internal/poolfile/config.go
@@ -119,7 +119,7 @@ func ProcessPool(poolFile *config.PoolFile, runnerName string) ([]drivers.Pool, 
 				google.WithDiskSize(g.Disk.Size),
 				google.WithDiskType(g.Disk.Type),
 				google.WithMachineImage(g.Image),
-				google.WithMachineType(g.MachineType),
+				google.WithSize(g.MachineType),
 				google.WithNetwork(g.Network),
 				google.WithSubnetwork(g.Subnetwork),
 				google.WithPrivateIP(g.PrivateIP),
